+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/write/BigQueryWriteHelper.java
@@ -146,13 +146,13 @@ public class BigQueryWriteHelper {
                         tableSchema,
                         config.getEnableModeCheckForSchemaFields())
                     .getTableId());
-        loadDataToBigQuery();
+        loadDataToBigQuery(/* applyDestinationTableLayoutOptions= */ false);
         Job queryJob =
             bigQueryClient.overwriteDestinationWithTemporaryDynamicPartitons(
                 temporaryTableId.get(), config.getTableId());
         bigQueryClient.waitForJob(queryJob, JobOperation.DYNAMIC_PARTITION_OVERWRITE);
       } else {
-        loadDataToBigQuery();
+        loadDataToBigQuery(/* applyDestinationTableLayoutOptions= */ true);
       }
       updateMetadataIfNeeded();
     } catch (Exception e) {
@@ -162,7 +162,7 @@ public class BigQueryWriteHelper {
     }
   }
 
-  void loadDataToBigQuery() throws IOException {
+  void loadDataToBigQuery(boolean applyDestinationTableLayoutOptions) throws IOException {
     FileSystem fs = gcsPath.getFileSystem(conf);
     FormatOptions formatOptions = config.getIntermediateFormat().getFormatOptions();
     String suffix = "." + formatOptions.getType().toLowerCase();
@@ -184,7 +184,8 @@ public class BigQueryWriteHelper {
             formatOptions,
             writeDisposition,
             Optional.of(tableSchema),
-            destinationTableId);
+            destinationTableId,
+            applyDestinationTableLayoutOptions);
 
     long currentTimeMillis = System.currentTimeMillis();
     SparkBigQueryConnectorMetricsUtils.postWriteSessionMetrics(